Move the dynamic dictionary generation into the dictionary-plugin
- Use `http.get()` to get the dictionary template and the wlking ctrller
  port data structure.
- Call the ditionary generator function from the `getDictionary()`.
