+++ openmctStaticServer/index.html
@@ -5,6 +5,7 @@
     <script src="openmctdist/openmct.js"></script>
     <script src="lib/http.js"></script>
     <script src="namespacesNtypes.js"></script>
+    <script src="genDictFromWalkingCtrlPortDataStruct.js"></script>
     <script src="dictionary-plugin.js"></script>
     <script src="historical-telemetry-plugin.js"></script>
     <script src="realtime-telemetry-plugin.js"></script>
